Enable Primo Victoria and Verdun options by date
Added logic to display and enable the 'Primo Victoria' option on June 6 and the 'Fields of Verdun' option between February 21 and December 18. Updated HTML to add an ID and disabled attribute to the Primo Victoria option, and adjusted JavaScript to handle their visibility and state based on the current date.